move to another part of the file and clean the parameter
