+++ Asterix/optics/propagation_functions.py
@@ -1,5 +1,5 @@
 import numpy as np
-from Asterix.utils import crop_or_pad_image, save_plane_in_fits
+from Asterix.utils import crop_or_pad_image
 
 
 def mft(image,
@@ -512,241 +512,3 @@ def fft_choosecenter(image, inverse=False, center_pos='bb', norm='backward'):
     farray *= np.exp(sens * (2. * 1j * np.pi / np.sqrt(Nx * Ny)) * np.sum(direct * fourier))
 
     return farray
-
-
-def butterworth_circle(dim, size_filter, order=5, xshift=0, yshift=0):
-    """
-    Return a circular Butterworth filter.
-
-    AUTHOR: Raphaël Galicher (in IDL)
-            ILa (to Python)
-
-    Parameters
-    ----------
-    dim : int
-        Dimension of 2D output array in pixels. If even, filter will be centered on a pixel, but can be shifted to
-        between pixels by using xshift=-0.5 and yshift=-0.5. If uneven, filter will be centered between pixels.
-    size_filter : int
-        Inverse size of the filter.
-    order : int
-        Order of the filter.
-    xshift : float
-        Shift of filter with respect to its array in the x direction, in pixels.
-    yshift : float
-        Shift of filter with respect to its array in the y direction, in pixels.
-
-    Returns
-    -------
-    butterworth : 2D array
-
-    Example
-    --------
-    siz = 100
-    rad = int(siz / 2)
-
-    bfilter3 = butterworth_circle(siz, rad, order=3, xshift=-0.5, yshift=-0.5)
-    bfilter5 = butterworth_circle(siz, rad, order=5, xshift=-0.5, yshift=-0.5)
-
-    plt.figure(figsize=(16, 8))
-    plt.subplot(1,2,1)
-    plt.imshow(bfilter5, cmap='Greys_r', origin='lower')
-    plt.title("Order = 5")
-    plt.colorbar()
-    plt.subplot(1,2,2)
-    plt.plot(bfilter3[rad], label="order=3")
-    plt.plot(bfilter5[rad], label="order=5")
-    plt.legend()
-    plt.show()
-    """
-    ty = (np.arange(dim) - yshift - dim / 2)
-    tx = (np.arange(dim) - xshift - dim / 2)
-    xx, yy = np.meshgrid(ty, tx)
-
-    butterworth = 1 / np.sqrt(1 + (np.sqrt(xx**2 + yy**2) / np.abs(size_filter) * 2)**(2. * order))
-
-    return butterworth
-
-
-def prop_fpm_regional_sampling(pup,
-                               fpm,
-                               real_dim_input=None,
-                               nbres=[4, 50],
-                               shift=(0, 0),
-                               filter_order=15,
-                               alpha=1.5,
-                               dir_save_all_planes=None):
-    """
-    Calculate the coronagraphic electric field in the Lyot plane by using varying sampling in different parts of the FPM.
-
-    Starting from the electric field in an input pupil plane, propagate through a phase-mask coronagraph with
-    different sampling values in different rings around the FPM center. Each propagation goes from the pupil to the FPM
-    and then to the Lyot plane (without Lyot stop applied), using MFTs, while a (combination of) Butterworth filters
-    marks annular regions around the FPM center that are considered in the propagation in each step. The electric field
-    contribution from each step is then summed up to yield the total electric field in the Lyot plane (before the LS).
-
-    AUTHOR: R. Galicher (in IDL)
-            ILa (to Python)
-            J. Mazoyer : small modifs to adapt to Asterix Dec 22
-
-    Parameters
-    ----------
-    pup : 2D array of size [self.dim_overpad_pupil, self.dim_overpad_pupil]
-        Input mage array containing the wavefront at the entrance pupil of the optical system.
-    fpm : 2D array
-        Complex electric field in the focal plane of the focal-plane mask.
-    real_dim_input : int or None, default None
-        Diameter of the support in pup (can differ from pup.shape). If None, real_dim_input = pup.shape
-        If the pupil fills its array out until the edges, then real_dim_input = pup.shape; if the pupil
-        is padded, then real_dim_input is the size of the pupil in pixels.
-    nbres : list
-        List of the number of resolution elements across the total image plane for all propagation layers.
-        As a general rule, it is probably safest to choose these numbers such that there is not sampling
-        shift right in the middle of the DH. This would ensure that the frequencies inside the DH are all
-        calculated with the same resolution in the EFC matrix. So if we correct the DH between two radii
-        IWA and OWA (in lambda/D), nbres should not have any elements between 2IWA and 2OWA.
-    shift : tuple, default (0, 0)
-        Shift of FPM with respect to optical axis in units of lambda/D.
-    filter_order : int
-        Order of the Butterworth filter.
-    alpha : float
-        Scale factor for the filter size. The larger this number, the smaller the filter size with respect to the
-        input array.
-    dir_save_all_planes : string or None, default None
-        If not None, absolute directory to save all planes in fits for debugging purposes.
-        This can generate a lot of fits especially if in a loop, use with caution.
-
-    Returns
-    -------
-    EF_before_LS : 2D array (complex)
-        E-field before the Lyot stop.
-    """
-    samp_outer = 2
-
-    if not isinstance(nbres, list):
-        raise TypeError(f"'nbres' parameter needs to be of type list or numpy.array. Currently type = {type(nbres)}")
-
-    nbres = np.array(nbres)
-
-    dim_overpad_pupil = pup.shape[0]
-    dim_fpm = fpm.shape[0]
-
-    if real_dim_input is None:
-        real_dim_input = dim_overpad_pupil
-    samplings = real_dim_input / nbres
-
-    if not np.all(np.diff(nbres) >= 0):
-        raise ValueError(f"'nbres' parameter needs to be sorted from the highest to lowest number of elements."
-                         f"Currently 'nbres' = {nbres}")
-
-    if np.min(samplings) < 2:
-        raise ValueError(f"The outer sampling in prop_fpm_regional_sampling is hardcoded to 2, otherwise we cut off"
-                         f"the high-spatial frequencies and the simulation turns out bad. We need the samplings"
-                         f"defined by the 'nbres' parameter (real_dim_input/nbres) to be always >= 2. Currently, with"
-                         f"real_dim_input = {real_dim_input}, samplings are {samplings}.")
-
-    if np.min(samplings) == 2:
-        # The outer sampling is already 2 (samp_outer = 2). In the case where the maximum
-        # element in 'nbres' corresponds to a sampling of 2, we can remove it and gain some time.
-        nbres = np.delete(nbres, -1)
-        samplings = np.delete(samplings, -1)
-
-    if max(shift) >= nbres[0] / 2:
-        raise ValueError(f"shift {shift} is larger than the minimum number of elements of resolution {nbres[0]} / 2:"
-                         f"the tip/tilt is out of the array! Increase min(nbres) or decrease shift.")
-
-
-    # Innermost part of the focal plane
-    but_inner = butterworth_circle(dim_fpm, dim_fpm / alpha, filter_order, xshift=-0.5, yshift=-0.5)
-    efield_before_fpm_inner = mft(pup,
-                                  real_dim_input=real_dim_input,
-                                  dim_output=dim_fpm,
-                                  nbres=nbres[0],
-                                  norm='ortho',
-                                  X_offset_output=shift[0] * samplings[0],
-                                  Y_offset_output=shift[1] * samplings[0])
-    efield_before_ls = mft(efield_before_fpm_inner * fpm * but_inner,
-                           real_dim_input=dim_fpm,
-                           dim_output=real_dim_input,
-                           nbres=nbres[0],
-                           inverse=True,
-                           norm='ortho',
-                           X_offset_input=shift[0] * samplings[0],
-                           Y_offset_input=shift[1] * samplings[0])
-    if dir_save_all_planes is not None:
-        name_plane = f'FPbeforeFPM_nbr{int(nbres[0])}_sampling{int(samplings[0])}'
-        save_plane_in_fits(dir_save_all_planes, name_plane, efield_before_fpm_inner)
-        name_plane = f'FPafterButandFPM_nbr{int(nbres[0])}_sampling{int(samplings[0])}'
-        save_plane_in_fits(dir_save_all_planes, name_plane, efield_before_fpm_inner * fpm * but_inner)
-        name_plane = f'PPbeforeLyot_nbr{int(nbres[0])}_sampling{int(samplings[0])}'
-        save_plane_in_fits(dir_save_all_planes, name_plane, efield_before_ls)
-
-    # From inner to outer part of FPM
-    const_but = butterworth_circle(dim_fpm, dim_fpm / alpha, filter_order, xshift=-0.5, yshift=-0.5)
-    for k in range(nbres.shape[0] - 1):
-        # Butterworth filter in each layer
-        sizebut_here = dim_fpm / alpha * nbres[k] / nbres[k + 1]
-        but = (1 - butterworth_circle(dim_fpm, sizebut_here, filter_order, xshift=-0.5, yshift=-0.5)) * const_but
-
-        ef_pre_fpm = mft(
-            pup,
-            real_dim_input=real_dim_input,
-            dim_output=dim_fpm,
-            nbres=nbres[k + 1],
-            norm='ortho',
-            X_offset_output=shift[0] * samplings[k + 1],
-            Y_offset_output=shift[1] * samplings[k + 1],
-        )
-        ef_pre_ls = mft(ef_pre_fpm * fpm * but,
-                        real_dim_input=dim_fpm,
-                        dim_output=real_dim_input,
-                        nbres=nbres[k + 1],
-                        X_offset_input=shift[0] * samplings[k + 1],
-                        Y_offset_input=shift[1] * samplings[k + 1],
-                        norm='ortho',
-                        inverse=True)
-        if dir_save_all_planes is not None:
-            name_plane = f'FPbeforeFPM_nbr{int(nbres[k + 1])}_sampling{int(samplings[k + 1])}'
-            save_plane_in_fits(dir_save_all_planes, name_plane, ef_pre_fpm)
-            name_plane = f'FPafterButandFPM_nbr{int(nbres[k + 1])}_sampling{int(samplings[k + 1])}'
-            save_plane_in_fits(dir_save_all_planes, name_plane, ef_pre_fpm * fpm * but)
-            name_plane = f'PPbeforeLyot_nbr{int(nbres[k + 1])}_sampling{int(samplings[k + 1])}'
-            save_plane_in_fits(dir_save_all_planes, name_plane, ef_pre_ls)
-
-        # Sum up E-field contributions before the LS
-        efield_before_ls += ef_pre_ls
-
-    # Outer part of the FPM
-    nbres_outer = dim_fpm / samp_outer
-    sizebut_outer = dim_fpm / alpha * nbres[-1] / nbres_outer
-    but_outer = 1 - butterworth_circle(dim_fpm, sizebut_outer, filter_order, xshift=-0.5, yshift=-0.5)
-
-    ef_pre_fpm_outer = mft(pup,
-                           real_dim_input=real_dim_input,
-                           dim_output=dim_fpm,
-                           nbres=nbres_outer,
-                           norm='ortho',
-                           X_offset_output=shift[0] * samp_outer,
-                           Y_offset_output=shift[1] * samp_outer)
-    ef_pre_ls_outer = mft(ef_pre_fpm_outer * fpm * but_outer,
-                          real_dim_input=dim_fpm,
-                          dim_output=real_dim_input,
-                          nbres=nbres_outer,
-                          norm='ortho',
-                          inverse=True,
-                          X_offset_input=shift[0] * samp_outer,
-                          Y_offset_input=shift[1] * samp_outer)
-
-    if dir_save_all_planes is not None:
-        name_plane = f'FPbeforeFPM_nbr{int(nbres_outer)}_sampling{int(samp_outer)}'
-        save_plane_in_fits(dir_save_all_planes, name_plane, ef_pre_fpm_outer)
-        name_plane = f'FPafterButandFPM_nbr{int(nbres_outer)}_sampling{int(samp_outer)}'
-        save_plane_in_fits(dir_save_all_planes, name_plane, ef_pre_fpm_outer * fpm * but_outer)
-        name_plane = f'PPbeforeLyot_nbr{int(nbres_outer)}_sampling{int(samp_outer)}'
-        save_plane_in_fits(dir_save_all_planes, name_plane, ef_pre_ls_outer)
-
-    # Total E-field before the LS
-    efield_before_ls += ef_pre_ls_outer
-
-    efield_before_ls = crop_or_pad_image(efield_before_ls, dim_overpad_pupil)
-
-    return efield_before_ls
