+++ Asterix/tests/param_file_tests.ini
@@ -22,7 +22,7 @@ onbench=False
     nb_wav = 15
 
     #Dimension of science image
-    dimScience = 400  #detector science image size (in pixels)
+    dimScience = 200  #detector science image size (in pixels)
 
     # Sampling in the detector science image
     #(lambda/Entrance_pupil_diameter in pixel
